requirements.txt
attrs==21.2.0
boto==2.49.0
boto3==1.20.2
botocore==1.23.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cmake==3.26.3
faiss-gpu==1.7.2
filelock==3.12.0
findspark==1.4.2
Flask==2.0.2
Flask-Cors==3.0.10
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.2.0
lit==16.0.3
MarkupSafe==2.0.1
mod-wsgi==4.9.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==21.2
pandas==2.0.1
pluggy==1.0.0
py==1.11.0
py4j==0.10.9.2
pyparsing==2.4.7
pyspark==3.2.0
pytest==6.2.5
python-dateutil==2.8.2
python-devtools==2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
s3transfer==0.5.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==5.2.1
sympy==1.12
threadpoolctl==3.0.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
tqdm==4.62.3
transformers==4.29.2
triton==2.0.0
typing-extensions==3.10.0.2
tzdata==2023.3
urllib3==1.26.7
watchtower==1.0.6
Werkzeug==2.0.2
zipp==3.6.0
